% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptative_arima.R
\name{adaptative_arima}
\alias{adaptative_arima}
\title{Adaptative Arima}
\usage{
adaptative_arima(
  model,
  data_train = Data0[sel_b, ],
  data_test = Data1,
  update_frequency = 20,
  rolling_window = 400
)
}
\arguments{
\item{model}{the model on which we will apply an ARIMA correction on the residuals}

\item{data_train}{the initial dataset on which we adjust an ARIMA model}

\item{data_test}{the dataset used to update the ARIMA model}

\item{update_frequency}{the number of passed days before readjusting the ARIMA model}

\item{rolling_window}{the length of the residuals vector with which the ARIMA will be adjusted}
}
\value{
a vector of 'Load' predictions based on forecasts obtained with the considered model
corrected with the ARIMA sequentially fitted
}
\description{
adjust an initial ARIMA model based on training data, then updates the ARIMA model
on a regular time basis based on previously observed residuals in test data
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
