% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptative_linear_model.R
\name{adaptative_linear_model}
\alias{adaptative_linear_model}
\title{Adaptative linear model}
\usage{
adaptative_linear_model(
  data_train,
  data_test,
  update_frequency = 50,
  rolling_window = 500
)
}
\arguments{
\item{data_train}{the training set}

\item{data_test}{the test set}

\item{update_frequency}{the size of the time window in days before a model update}

\item{rolling_window}{the size of the time window in days used to train the models}
}
\value{
a vector containing the forecasts obtained with the adjusted models
}
\description{
Adjust sequentially a linear model based on feature combinations and AIC penalty method
The model is trained on a rolling-window and updated including past residuals observed on test set
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
