% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptative_linear_model_AIC.R
\name{adaptative_linear_model_AIC}
\alias{adaptative_linear_model_AIC}
\title{Adaptative Linear model with AIC penalty and forgetting factor}
\usage{
adaptative_linear_model_AIC(
  data_train,
  data_test,
  update_frequency = 50,
  rolling_window = 1500
)
}
\arguments{
\item{data_train}{the train set}

\item{data_test}{the test set}

\item{update_frequency}{the number of days before refitting a linear model}

\item{rolling_window}{the number of days kept in memory for model training}
}
\value{
the forecast obtained
}
\description{
Refit linear models through time based on a rolling window of last observations
and penalizing with sigmoid weights the oldest training examples
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
