% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptative_rf_model.R
\name{adaptative_rf_model}
\alias{adaptative_rf_model}
\title{Adaptative Random Forest model}
\usage{
adaptative_rf_model(
  train = Data0[sel_a, ],
  covid_period = Data0[sel_b, ],
  test = Data1,
  trees = 10,
  var_allowed = 7,
  rolling_window = 2000
)
}
\arguments{
\item{train}{the data on which we will fit the initial model}

\item{covid_period}{the subset of train test during COVID}

\item{test}{the test set}

\item{trees}{the number of trees in each forest}

\item{var_allowed}{the number of variables to possibly split at each node}
}
\value{
the forecast on the train COVID period and the test set
}
\description{
Online learning of Random Forests
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
