% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockCV.R
\name{build_block}
\alias{build_block}
\title{Build_block}
\usage{
build_block(dataset = Data0[sel_a], Nblock = 8)
}
\arguments{
\item{dataset}{the dataset used for cross-validation}

\item{Nblock}{the number of folds used}
}
\description{
Build blocks of data prior to leave-one-out cross-validation
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
