% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{feature_engineering}
\alias{feature_engineering}
\title{Feature Engineering}
\usage{
feature_engineering(dataset)
}
\arguments{
\item{dataset}{the dataset on which the transformations are applied}
}
\value{
the modified dataset
}
\description{
Operates diverse basic transformations on datasets
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
