% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockCV.R
\name{fit_gam_mod}
\alias{fit_gam_mod}
\title{GAM on block}
\usage{
fit_gam_mod(equation, block)
}
\arguments{
\item{equation}{the equation for the model}

\item{block}{a vector with the indexes of the row used for CV}
}
\description{
Adjust GAM on block for CV
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
