% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockCV.R
\name{fit_lm_mod}
\alias{fit_lm_mod}
\title{Linear model on block}
\usage{
fit_lm_mod(eq, block)
}
\arguments{
\item{eq}{the equation for the model}

\item{block}{a vector with the indexes of the row used for CV}
}
\description{
Adjust linear model on block for CV
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
