% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockCV.R
\name{gam_cvpred}
\alias{gam_cvpred}
\title{GAM block Cross-Validation}
\usage{
gam_cvpred(dataset = Data0[sel_a, ], equation)
}
\arguments{
\item{dataset}{train set}

\item{equation}{the equation used for the model}
}
\description{
Leave-one-out K-fold cross-validation to evaluate GAM performance on the training set subset anterior to COVID
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
