% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_lm_model.R
\name{gam_lm_model}
\alias{gam_lm_model}
\title{GAM + LM}
\usage{
gam_lm_model(dataset = Data0[sel_a, ], weighted = T)
}
\arguments{
\item{weighted}{logical used to determine if we fit a time-weighted GAM or not. If true, use exponentially decreasing weights.}

\item{model}{the dataset for training}
}
\value{
the fitted model
}
\description{
Fit a Generalized Additive model, time-weighted or not, adjusted with features constructed for the linear model
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
