% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuned_gam_model.R
\name{gam_tuned_model}
\alias{gam_tuned_model}
\title{Tuned GAM}
\usage{
gam_tuned_model(dataset = Data0[sel_a, ])
}
\arguments{
\item{dataset}{the data on which we will fit the model}
}
\value{
the fitted GAM
}
\description{
adjust a GAM based on tuned equation
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
