% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_model.R
\name{gbm_model}
\alias{gbm_model}
\title{Generalized Boosted Regresison Modeling}
\usage{
gbm_model(dataset = Data0[sel_a, ], distribution)
}
\arguments{
\item{dataset}{the dataset on which the model is fitted}
}
\value{
the selected model
}
\description{
Fit a linear model to the data, constructing polynomial combinations of the features
and choosing the proper model by AIC in bidirectional stepwise algorithm
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
