% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_significance.R
\name{interaction_significance}
\alias{interaction_significance}
\title{Interaction significance}
\usage{
interaction_significance(dataset = Data0[sel_a, ], feature1, feature2)
}
\arguments{
\item{dataset}{the data on which the models are fitted}

\item{feature1}{the first feature to test}

\item{feature2}{the second feature to test}
}
\value{
the p-value of a Chi-squared test on slotted models
}
\description{
Perform a Chi-squared test between slotted models to assess the significance
of an interaction between two features
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
