% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_AIC.R
\name{linear_model_AIC}
\alias{linear_model_AIC}
\title{Linear model with AIC penalty}
\usage{
linear_model_AIC(dataset = Data0[sel_a, ])
}
\arguments{
\item{dataset}{the dataset on which the model is fitted}
}
\value{
the selected model
}
\description{
Fit a linear model to the data, constructing polynomial combinations of the features
and choosing the proper model by AIC in bidirectional stepwise algorithm
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
