% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockCV.R
\name{lm_cvpred}
\alias{lm_cvpred}
\title{Linear model block Cross-Validation}
\usage{
lm_cvpred(dataset = Data0[sel_a, ], eq)
}
\arguments{
\item{dataset}{train set}

\item{eq}{the equation used for the model}
}
\description{
Leave-one-out K-fold cross-validation to evaluate LM performance on the training set subset anterior to COVID
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
