% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_model.R
\name{monthly_model}
\alias{monthly_model}
\title{Monthly linear model with AIC penalty}
\usage{
monthly_model(
  dataset = Data0[sel_a, ],
  covid_period = rbind(Data0[sel_b, -2], Data1[, -20])
)
}
\arguments{
\item{dataset}{the dataset on which the models are fitted}

\item{covid_period}{the subset of train set during COVID and the test set}
}
\value{
the forecasts computed with the different models on COVID period
}
\description{
Fit a linear model for each month based on the equation found with AIC penalty
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
