% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_qgam_model.R
\name{naive_qgam_model}
\alias{naive_qgam_model}
\title{Naive QGAM}
\usage{
naive_qgam_model(dataset = Data0[sel_a, ], quantile = 0.3)
}
\arguments{
\item{dataset}{the dataset on which the model is fitted}

\item{quantile}{the quantile selected}
}
\value{
the selected model
}
\description{
Fit a smooth additive quantile regression model based on a light equation
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
