% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_qgam_residual_variance.R
\name{naive_qgam_residual_variance}
\alias{naive_qgam_residual_variance}
\title{Naive QGAM residual variance}
\usage{
naive_qgam_residual_variance(dataset = Data0[sel_a, ], model)
}
\arguments{
\item{dataset}{the dataset on which the residuals are observed}

\item{quantile}{the naive QGAM}
}
\value{
the obtained residuals and variance plots by temperature levels
}
\description{
Observe the residual variance obtained with the naive QGAM and compare it based on temperature levels
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
