% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refined_qgam.R
\name{refined_qgam_model}
\alias{refined_qgam_model}
\title{Refined QGAM}
\usage{
refined_qgam_model(dataset = Data0[sel_a, ], quantile = 0.3)
}
\arguments{
\item{dataset}{the dataset on which the model is fitted}
}
\value{
the selected model
}
\description{
Fit a smooth additive quantile regression model based on a light equation and
with an adaptative learning rate based
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
