% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_variable_selection.R
\name{rf_variable_selection}
\alias{rf_variable_selection}
\title{Variable selection with Random Forest}
\usage{
rf_variable_selection(dataset = Data0[sel_a, ], var_allowed = 13, trees = 200)
}
\arguments{
\item{dataset}{the data on which we will fit the model}

\item{var_allowed}{the number of variables to possibly split at each node}

\item{trees}{the number of trees in the forest}
}
\value{
the vector of variable importance
}
\description{
Adjust a random forest so as to visualize
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
