% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error}
\usage{
rmse(y, yhat, digits = 2)
}
\arguments{
\item{y}{the observations to be predicted}

\item{yhat}{the predictions}

\item{digits}{the precision in number of digits}
}
\value{
a positive real number the RMSE
}
\description{
compute the Root Mean Square Error between a vector y and its forecast yhat
}
\examples{
y<-rnorm(10)
yhat<-rep(0,10)
rmse(y,yhat,digits=4)
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
