% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_model.R
\name{xgb_linear_model}
\alias{xgb_linear_model}
\title{XGBoost Linear model}
\usage{
xgb_linear_model(dataset = Data0[sel_a, ])
}
\arguments{
\item{dataset}{the dataset on which the model is fitted}
}
\value{
the fitted model
}
\description{
Fit a XGBoost linear model
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
