% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_model.R
\name{xgb_predict}
\alias{xgb_predict}
\title{XGBoost predict}
\usage{
xgb_predict(model = xgb_model, covid_period = Data0[sel_b, ], test = Data1)
}
\arguments{
\item{model}{the XGBoost model trained}

\item{covid_period}{the COVID subset of train set}

\item{test}{the test set}
}
\value{
a vector of forecasts
}
\description{
Make predictions on COVID subset of train set and test set based on
XGBoost model trained
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
