% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_model.R
\name{xgb_tree_model}
\alias{xgb_tree_model}
\title{XGBoost Tree model}
\usage{
xgb_tree_model(dataset = Data0[sel_a, ])
}
\arguments{
\item{dataset}{the dataset on which the model is fitted}
}
\value{
the fitted model
}
\description{
Fit a XGBoost tree model
}
\author{
halvard.bariller@universite-paris-saclay.fr>
}
